+++ internal/airlock/airlock.go
@@ -22,16 +22,15 @@ import (
 	"github.com/neicnordic/crypt4gh/keys"
 	"github.com/neicnordic/crypt4gh/model/headers"
 	"github.com/neicnordic/crypt4gh/streaming"
+	"golang.org/x/crypto/chacha20poly1305"
 )
 
 var ai airlockInfo = airlockInfo{}
 var infoFile = "/etc/pam_userinfo/config.json"
 var minimumSegmentSize = 1 << 20
 
-const keySize = 32
-
 type airlockInfo struct {
-	publicKey [keySize]byte
+	publicKey [chacha20poly1305.KeySize]byte
 	proxy     string
 	project   string
 }
@@ -132,7 +131,7 @@ func GetPublicKey() error {
 	if err != nil {
 		return fmt.Errorf("Could not decode public key: %w", err)
 	}
-	if len(data) < keySize {
+	if len(data) < chacha20poly1305.KeySize {
 		return fmt.Errorf("Invalid length of decoded public key (%v)", len(data))
 	}
 
@@ -249,8 +248,10 @@ func Upload(original_filename, filename, container, journal_number string, segme
 		return nil
 	}
 
-	err = <-errc
-	return fmt.Errorf("Streaming file failed: %w", err)
+	if err = <-errc; err != nil {
+		return fmt.Errorf("Streaming file failed: %w", err)
+	}
+	return nil
 }
 
 func reorderNames(filename, directory string) (string, string) {
@@ -296,7 +297,9 @@ var getFileDetails = func(filename string) (*os.File, string, int64, error) {
 }
 
 var newCrypt4GHWriter = func(w io.Writer) (io.WriteCloser, error) {
-	return streaming.NewCrypt4GHWriterWithoutPrivateKey(w, ai.publicKey, nil)
+	pubkeyList := [][chacha20poly1305.KeySize]byte{}
+	pubkeyList = append(pubkeyList, ai.publicKey)
+	return streaming.NewCrypt4GHWriterWithoutPrivateKey(w, pubkeyList, nil)
 }
 
 var encrypt = func(file *os.File, pw *io.PipeWriter, errc chan error) {