+++ CHANGELOG.md
@@ -3,7 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-## Unreleased
+
+## [Unreleased]
+
+### Changed
+- `crypt4gh` function `NewCrypt4GHWriterWithoutPrivateKey` now uses list of public keys
 
 ### Added
 - Option to input airlock password from environment variable